requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
Pillow>=10.4.0
